+++ front/components/assistant/manager/AssistantsTable.tsx
@@ -1,26 +1,11 @@
 import {
   Avatar,
   BracesIcon,
-  Button,
   Checkbox,
-  ChevronDownIcon,
   Chip,
   ClipboardIcon,
-  Cog6ToothIcon,
   DataTable,
-  Dialog,
-  DialogContainer,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
   PencilSquareIcon,
-  SliderToggle,
-  Spinner,
   Tooltip,
   TrashIcon,
 } from "@dust-tt/sparkle";
@@ -30,8 +15,8 @@ import { useMemo, useState } from "react";
 
 import { DeleteAssistantDialog } from "@app/components/assistant/DeleteAssistantDialog";
 import { assistantUsageMessage } from "@app/components/assistant/Usage";
-import { useTheme } from "@app/components/sparkle/ThemeContext";
-import { useTags, useUpdateAgentTags } from "@app/lib/swr/tags";
+import { SCOPE_INFO } from "@app/components/assistant_builder/Sharing";
+import { useTags } from "@app/lib/swr/tags";
 import { classNames, formatTimestampToFriendlyDate } from "@app/lib/utils";
 import type {
   AgentConfigurationScope,
@@ -39,80 +24,10 @@ import type {
   LightAgentConfigurationType,
   WorkspaceType,
 } from "@app/types";
-import { isAdmin, isBuilder, pluralize } from "@app/types";
+import { isAdmin, pluralize } from "@app/types";
 import type { TagType } from "@app/types/tag";
-import { SCOPE_INFO } from "@app/components/assistant_builder/Sharing";
-
-const TagSelector = ({
-  tags,
-  agentTags,
-  agentConfigurationId,
-  owner,
-  onChange,
-}: {
-  tags: TagType[];
-  agentTags: TagType[];
-  agentConfigurationId: string;
-  owner: WorkspaceType;
-  onChange: () => Promise<any>;
-}) => {
-  const [isLoading, setIsLoading] = useState(false);
-  const { isDark } = useTheme();
-  const updateAgentTags = useUpdateAgentTags({
-    owner,
-    agentConfigurationId,
-  });
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger>
-        {agentTags.length === 0 ? (
-          <Button
-            variant="ghost"
-            icon={ChevronDownIcon}
-            size="xs"
-            label="Add tags"
-            className="invisible text-muted-foreground group-hover:visible"
-          />
-        ) : (
-          <Button
-            variant="ghost"
-            icon={ChevronDownIcon}
-            size="xmini"
-            className="invisible text-muted-foreground group-hover:visible"
-          />
-        )}
-      </DropdownMenuTrigger>
-      <DropdownMenuContent>
-        {tags.map((t) => {
-          const isChecked = agentTags.some((x) => x.sId === t.sId);
-          return (
-            <DropdownMenuCheckboxItem
-              key={t.sId}
-              onClick={async (e: React.MouseEvent) => {
-                setIsLoading(true);
-                e.stopPropagation();
-                e.preventDefault();
-                await updateAgentTags({
-                  addTagIds: isChecked ? [] : [t.sId],
-                  removeTagIds: isChecked ? [t.sId] : [],
-                });
-                await onChange();
-              }}
-              checked={isChecked}
-            >
-              <Chip size="xs" label={t.name} color="golden" />
-            </DropdownMenuCheckboxItem>
-          );
-        })}
-        {isLoading && (
-          <div className="absolute inset-0 flex items-center justify-center bg-white/50 dark:bg-black/50">
-            <Spinner variant={isDark ? "light" : "dark"} />
-          </div>
-        )}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-};
+import { TableTagSelector } from "@app/components/assistant/manager/TableTagSelector";
+import { GlobalAgentAction } from "@app/components/assistant/manager/GlobalAgentAction";
 
 type MoreMenuItem = {
   label: string;
@@ -230,7 +145,7 @@ const getTableColumns = ({
                       trigger={<span>{info.getValue()}</span>}
                     />
                   </div>
-                  <TagSelector
+                  <TableTagSelector
                     tags={tags}
                     agentTags={info.row.original.agentTags}
                     agentConfigurationId={info.row.original.sId}
@@ -325,94 +240,7 @@ const getTableColumns = ({
   ];
 };
 
-type GlobalAgentActionProps = {
-  agent: LightAgentConfigurationType;
-  owner: WorkspaceType;
-  handleToggleAgentStatus: (
-    agent: LightAgentConfigurationType
-  ) => Promise<void>;
-  showDisabledFreeWorkspacePopup: string | null;
-  setShowDisabledFreeWorkspacePopup: (s: string | null) => void;
-};
-
-function GlobalAgentAction({
-  agent,
-  owner,
-  handleToggleAgentStatus,
-  showDisabledFreeWorkspacePopup,
-  setShowDisabledFreeWorkspacePopup,
-}: GlobalAgentActionProps) {
-  const router = useRouter();
-  if (agent.sId === "helper") {
-    return null;
-  }
-
-  if (agent.sId === "dust") {
-    return (
-      <Button
-        variant="outline"
-        icon={Cog6ToothIcon}
-        size="xs"
-        disabled={!isBuilder(owner)}
-        onClick={(e: Event) => {
-          e.stopPropagation();
-          void router.push(`/w/${owner.sId}/builder/assistants/dust`);
-        }}
-      />
-    );
-  }
-
-  return (
-    <>
-      <SliderToggle
-        size="xs"
-        onClick={async (e) => {
-          e.stopPropagation();
-          await handleToggleAgentStatus(agent);
-        }}
-        selected={agent.status === "active"}
-        disabled={
-          !isBuilder(owner) || agent.status === "disabled_missing_datasource"
-        }
-      />
-      <div className="whitespace-normal" onClick={(e) => e.stopPropagation()}>
-        <Dialog
-          open={showDisabledFreeWorkspacePopup === agent.sId}
-          onOpenChange={(open) => {
-            if (!open) {
-              setShowDisabledFreeWorkspacePopup(null);
-            }
-          }}
-        >
-          <DialogContent size="md">
-            <DialogHeader hideButton={false}>
-              <DialogTitle>Free plan</DialogTitle>
-            </DialogHeader>
-            <DialogContainer>
-              {`@${agent.name} is only available on our paid plans.`}
-            </DialogContainer>
-            <DialogFooter
-              leftButtonProps={{
-                label: "Cancel",
-                variant: "outline",
-                onClick: () => setShowDisabledFreeWorkspacePopup(null),
-              }}
-              rightButtonProps={{
-                label: "Check Dust plans",
-                variant: "primary",
-                onClick: () => {
-                  void router.push(`/w/${owner.sId}/subscription`);
-                },
-              }}
-            />
-          </DialogContent>
-        </Dialog>
-      </div>
-    </>
-  );
-}
-
-type AgentsTableProps = {
+type AssistantsTableProps = {
   owner: WorkspaceType;
   agents: LightAgentConfigurationType[];
   setShowDetails: (agent: LightAgentConfigurationType) => void;
@@ -438,7 +266,7 @@ export function AssistantsTable({
   selection,
   setSelection,
   mutateAgentConfigurations,
-}: AgentsTableProps) {
+}: AssistantsTableProps) {
   const { tags } = useTags({ owner });
   const [showDeleteDialog, setShowDeleteDialog] = useState<{
     open: boolean;
